MainからControllerとServiceへの処理移管と各ファイル構成整理
