+++ src/main/java/raisetech/StudentManagement/StudentManagementApplication.java
@@ -1,38 +1,13 @@
 package raisetech.StudentManagement;
 
-import java.util.List;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
 
 @SpringBootApplication
-@RestController
 public class StudentManagementApplication {
-
-  @Autowired
-  public StudentRepository repository;
 
   public static void main(String[] args) {
     SpringApplication.run(StudentManagementApplication.class, args);
-  }
-
-  @GetMapping("/studentList")
-  public List<Student> getStudentList() {
-    return repository.searchStudentList();
-  }
-
-  //todo:引数をpublicIdに変更する
-  @GetMapping("/student")
-  public Student getStudentById(@RequestParam Integer studentId) {
-    return repository.searchStudentById(studentId);
-  }
-
-  @GetMapping("/studentCourseList")
-  public List<StudentCourse> getStudentCourseList() {
-    return repository.searchStudentCourseList();
   }
 
 }
