+++ src/main/java/raisetech/StudentManagement/controller/StudentController.java
@@ -1,0 +1,38 @@
+package raisetech.StudentManagement.controller;
+
+import java.util.List;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+import raisetech.StudentManagement.date.Student;
+import raisetech.StudentManagement.date.StudentCourse;
+import raisetech.StudentManagement.service.StudentService;
+
+@RestController
+public class StudentController {
+
+  public StudentService service;
+
+  @Autowired
+  public StudentController(StudentService service) {
+    this.service = service;
+  }
+
+  @GetMapping("/studentList")
+  public List<Student> getStudentList() {
+    return service.searchStudentList();
+  }
+
+  //todo:引数をpublicIdに変更する
+  @GetMapping("/student")
+  public Student getStudentById(@RequestParam Integer studentId) {
+    return service.searchStudentById(studentId);
+  }
+
+  @GetMapping("/studentCourseList")
+  public List<StudentCourse> getStudentCourseList() {
+    return service.searchStudentCourseList();
+  }
+
+}
