+++ src/main/java/raisetech/StudentManagement/service/StudentService.java
@@ -1,0 +1,34 @@
+package raisetech.StudentManagement.service;
+
+import java.util.List;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import raisetech.StudentManagement.date.Student;
+import raisetech.StudentManagement.date.StudentCourse;
+import raisetech.StudentManagement.repository.StudentRepository;
+
+@Service
+public class StudentService {
+
+  private StudentRepository repository;
+
+  @Autowired
+  public StudentService(StudentRepository repository) {
+    this.repository = repository;
+  }
+
+  public List<Student> searchStudentList() {
+    return repository.searchStudentList();
+  }
+
+  //todo:引数をpublicIdに変更する
+  public Student searchStudentById(Integer studentId) {
+    return repository.searchStudentById(studentId);
+  }
+
+  public List<StudentCourse> searchStudentCourseList() {
+    return repository.searchStudentCourseList();
+  }
+
+
+}
